% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglePoint.R
\name{anglePoint}
\alias{anglePoint}
\title{anglePoint function}
\usage{
anglePoint(d, theta, ref = c(0, 0))
}
\arguments{
\item{d}{numeric. distance traveled}

\item{theta}{numeric. angle in degrees to move in.}

\item{ref}{numeric.  Reference point to start at. length must be 2. Default is c(0, 0)}
}
\value{
numeric. point
}
\description{
function to find new point given a reference point, an angle and a distance.
}
\details{
shift point by angle and distance.
}
\examples{
anglePoint(d = 1, theta = 0, ref = c(0,0))
anglePoint(d = 2, theta = 90, ref = c(1,0))
}
