% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDist.R
\name{calcDist}
\alias{calcDist}
\title{calcDist function}
\usage{
calcDist(pt1, pt2, offset = c(5, 10), arlen = 0.1, printDist = TRUE,
  units = "ft", digits = NULL)
}
\arguments{
\item{pt1}{[value]}

\item{pt2}{[value]}

\item{offset}{[value]. Default is c(5, 10)}

\item{arlen}{[value]. Default is 0.10}

\item{printDist}{[value]. Default is TRUE}

\item{units}{[value]. Default is "ft"}

\item{digits}{[value]. Default is 0}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
