% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDistFromPlots.R
\name{calcDistFromPlots}
\alias{calcDistFromPlots}
\title{calcDistFromPlots function}
\usage{
calcDistFromPlots(plots1, plotNo1, corner1, plots2, plotNo2, corner2, ...)
}
\arguments{
\item{plots1}{[value]}

\item{plotNo1}{[value]}

\item{corner1}{[value]}

\item{plots2}{[value]}

\item{plotNo2}{[value]}

\item{corner2}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
