% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPal.R
\docType{data}
\name{cols10}
\alias{cols10}
\title{color palettes}
\format{An object of class \code{character} of length 10.}
\usage{
cols10
}
\description{
color palettes
}
\details{
various color palettes
}
\examples{
# none
}
\keyword{datasets}
