% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPal.R
\docType{data}
\name{cols12}
\alias{cols12}
\title{color pallete'12}
\format{
An object of class \code{character} of length 12.
}
\usage{
cols12
}
\description{
color pallete'12
}
\details{
color palette 12
}
\keyword{datasets}
