% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPal.R
\docType{data}
\name{cols16}
\alias{cols16}
\title{color pallete'16}
\format{
An object of class \code{character} of length 16.
}
\usage{
cols16
}
\description{
color pallete'16
}
\details{
color palette 16
}
\keyword{datasets}
