% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPal.R
\docType{data}
\name{cols20}
\alias{cols20}
\title{color pallete'20}
\format{
An object of class \code{character} of length 20.
}
\usage{
cols20
}
\description{
color pallete'20
}
\details{
color palette 20
}
\keyword{datasets}
