% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPal.R
\docType{data}
\name{cols30}
\alias{cols30}
\title{color pallete'30}
\format{
An object of class \code{character} of length 31.
}
\usage{
cols30
}
\description{
color pallete'30
}
\details{
color palette 30
}
\keyword{datasets}
