% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPal.R
\docType{data}
\name{cols32}
\alias{cols32}
\title{color pallete'32}
\format{
An object of class \code{character} of length 32.
}
\usage{
cols32
}
\description{
color pallete'32
}
\details{
color palette 32
}
\keyword{datasets}
