% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPal.R
\docType{data}
\name{cols40}
\alias{cols40}
\title{color pallete'40}
\format{An object of class \code{character} of length 40.}
\usage{
cols40
}
\description{
color pallete'40
}
\details{
color palette 40
}
\keyword{datasets}
