% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{combine function}
\usage{
combine(x1, x2, pile = FALSE)
}
\arguments{
\item{x1}{[value]}

\item{x2}{[value]}

\item{pile}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
