% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinePlots.R
\name{combinePlots}
\alias{combinePlots}
\title{combinePlots function}
\usage{
combinePlots(..., trialName = NULL, pile = NULL)
}
\arguments{
\item{...}{objects of class 'fieldPlots' to be combined.}

\item{trialName}{[value]. Default is NULL}

\item{pile}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
