% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compass.R
\name{compass}
\alias{compass}
\title{compass function}
\usage{
compass(x, y, angle, xsize = 0.5, ysize = 1, txtSize = 1,
  pointyness = 3, col = "black")
}
\arguments{
\item{x}{[value]}

\item{y}{[value]}

\item{angle}{[value]}

\item{xsize}{[value]. Default is 0.5}

\item{ysize}{[value]. Default is 1}

\item{txtSize}{[value]. Default is 1}

\item{pointyness}{[value]. Default is 3}

\item{col}{[value]. Default is "black"}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
