% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countFill.R
\name{countFill}
\alias{countFill}
\title{countFill function}
\usage{
countFill(l, fill = "fill")
}
\arguments{
\item{l}{[value]}

\item{fill}{[value]. Default is "fill"}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
