% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degToRad.R
\name{degToRad}
\alias{degToRad}
\title{degToRad function}
\usage{
degToRad(deg)
}
\arguments{
\item{degToRad}{<- function (deg [value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
