% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawLegend.R
\name{drawLegend}
\alias{drawLegend}
\title{drawLegend function}
\usage{
drawLegend(plotList, boundry, shiftLeg, cols, scaleLeg = 0.33,
  rotateLabs = 0, ...)
}
\arguments{
\item{plotList}{[value]}

\item{boundry}{[value]}

\item{shiftLeg}{[value]}

\item{cols}{[value]}

\item{scaleLeg}{[value]. Default is 0.33}

\item{rotateLabs}{[value]. Default is 0}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
