% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawPlots.R
\name{drawPlots}
\alias{drawPlots}
\title{drawPlots function}
\usage{
drawPlots(
  plots,
  angle,
  labAngle = 0,
  plotNos = TRUE,
  add = TRUE,
  ptrim = 1,
  rtrim = 2,
  color = NA,
  showBorderNum = FALSE,
  ...
)
}
\arguments{
\item{plots}{[value]}

\item{angle}{[value]}

\item{labAngle}{[value]. Default is 0}

\item{plotNos}{[value]. Default is TRUE}

\item{add}{[value]. Default is TRUE}

\item{ptrim}{[value]. Default is 1}

\item{rtrim}{[value]. Default is 2}

\item{color}{[value]. Default is NA}

\item{showBorderNum}{[value]. Default is FALSE}

\item{...}{arguments passed to text()}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
