% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldBlock.R
\docType{class}
\name{fieldBlock-class}
\alias{fieldBlock-class}
\alias{fieldBlock}
\title{An S4 class to represent field blocks, made up of field plots}
\description{
An S4 class to represent field blocks, made up of field plots
}
\section{Slots}{

\describe{
\item{\code{trial}}{matrix of trial}

\item{\code{plotNo}}{matrix of plotNo}

\item{\code{long}}{data.frame of long}

\item{\code{needStake}}{list of needStake}

\item{\code{borderPasses}}{numeric of borderPasses}

\item{\code{blockName}}{character of blockName}
}}

