% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldBoundary.R
\docType{class}
\name{fieldBoundary-class}
\alias{fieldBoundary-class}
\alias{fieldBoundary}
\title{An S4 class to represent field blocks, made up of field plots}
\description{
An S4 class to represent field blocks, made up of field plots
}
\section{Slots}{

\describe{
\item{\code{points}}{(n x 2) matrix of ordered field boundary points}

\item{\code{angle}}{numeric vector of n-1 of angles between points in degrees}

\item{\code{distance}}{distance between pairwise boundary points}

\item{\code{reference}}{nuermic vector of length 2, definingthe reference point}

\item{\code{borderPasses}}{numeric of borderPasses}

\item{\code{blockName}}{character of blockName}
}}

