% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldMap.R, R/fieldPlots.R
\docType{class}
\name{fieldPlots}
\alias{fieldPlots}
\alias{fieldMap}
\title{An S4 class to represent field plots}
\description{
An S4 class to represent field plots

An S4 class to represent field plots
}
\section{Slots}{

\describe{
\item{\code{blocks}}{list of blocks}

\item{\code{centers}}{list of centers}

\item{\code{corners}}{list of corners}

\item{\code{matrix}}{matrix of matrix}

\item{\code{needStake}}{list of needStake}

\item{\code{borderPasses}}{numeric of borderPasses}

\item{\code{trialName}}{character of trialName}
}}

