% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixPlantingErrors.R
\name{fixPlantingErrors}
\alias{fixPlantingErrors}
\title{fixPlantingMistakes function}
\usage{
fixPlantingErrors(plots, swap)
}
\arguments{
\item{plots}{object of class 'fieldPlots' or 'fieldBlock'}

\item{swap}{a list with each element of length 2, or matrix with two columns, where the first element or column is the plotName of the seed source that was planted into the space originally intended to be occupied by the plotName in the second element or column}
}
\value{
[value ] object of class 'fieldPlots' or 'fieldBlock' corresponding to 'plots' argument, with the Trial, Entry and Line swapped out.
}
\description{
function to fix planting mistakes in the field
}
\details{
[fill in details here]
}
\examples{
# none
}
