% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipInt.R
\name{flipInt}
\alias{flipInt}
\title{flipInt function}
\usage{
flipInt(x, maxx = NULL)
}
\arguments{
\item{x}{[value]}

\item{maxx}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to flip integers. 5:10 becomes 10:5.
}
\details{
[fill in details here]
}
\examples{
# none
}
