% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatScabBlock.R
\name{formatScabBlock}
\alias{formatScabBlock}
\title{formatScabBlock function to make scab trays.}
\usage{
formatScabBlock(block, trays, trim = NULL, flip = 1, ...)
}
\arguments{
\item{block}{object of class fieldBlock}

\item{trays}{data.frame with scab entries per tray.}

\item{trim}{numeric. Which ranges should be trimmed off of trays, typically the front (1) and last (20) ranges are planted in another species as border, so they need to be removed to format the scab block correctly.}

\item{flip}{default is 1}

\item{...}{additional arguments passed to makeTwoListMat()}
}
\value{
data.frame formated breedbase multi trial upload file
}
\description{
function to create scab information for blocks.
}
\details{
reformats blocks for the scab nursery
}
\examples{
# none.
}
