% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCenter.R
\name{getCenter}
\alias{getCenter}
\title{getCenter function}
\usage{
getCenter(plots)
}
\arguments{
\item{plots}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
