% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoord.R
\name{getCoord}
\alias{getCoord}
\title{getCoord function}
\usage{
getCoord(plots, plotNo, corner = "bottomright")
}
\arguments{
\item{plots}{[value]}

\item{plotNo}{[value]}

\item{corner}{[value]. Default is "bottomright"}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
