% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorner.R
\name{getCorner}
\alias{getCorner}
\title{getCorner function}
\usage{
getCorner(plots, whichCorner = "bottomright", whichPlotCorner = whichCorner)
}
\arguments{
\item{plots}{[value]}

\item{whichCorner}{[value]. Default is "bottomright"}

\item{whichPlotCorner}{[value]. Default is whichCorner}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
