% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGPStheta.R
\name{getGPStheta}
\alias{getGPStheta}
\title{getGPStheta function}
\usage{
getGPStheta(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{[value]}

\item{lon1}{[value]}

\item{lat2}{[value]}

\item{lon2}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
