% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutline.R
\name{getOutline}
\alias{getOutline}
\title{getOutline function}
\usage{
getOutline(plots)
}
\arguments{
\item{plots}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
