% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPStoDist.R
\name{GPStoDist}
\alias{GPStoDist}
\title{GPStoDist function}
\usage{
GPStoDist(boundry, units = "ft", refPoint = NULL)
}
\arguments{
\item{boundry}{[value]}

\item{units}{[value]. Default is "ft"}

\item{refPoint}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
