% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvestBags.R
\docType{class}
\name{harvestBags-class}
\alias{harvestBags-class}
\alias{harvestBags}
\title{An S4 class to represent information to be printed on harvest bags}
\description{
An S4 class to represent information to be printed on harvest bags
}
\section{Slots}{

\describe{
\item{\code{plotNo}}{character of plotNo}

\item{\code{pass}}{numeric of pass}

\item{\code{range}}{numeric of range}

\item{\code{trial}}{character of trial}

\item{\code{block}}{character of block}

\item{\code{long}}{data.frame of long}
}}

