% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haversine.R
\name{haversine}
\alias{haversine}
\title{haversine function}
\usage{
haversine(lat1, lon1, lat2, lon2, r = 6371, units = "km")
}
\arguments{
\item{lat1}{[value]}

\item{lon1}{[value]}

\item{lat2}{[value]}

\item{lon2}{[value]}

\item{r}{[value]. Default is 6371}

\item{units}{[value]. Default is "km"}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
