% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelPasses.R
\name{labelPasses}
\alias{labelPasses}
\title{labelPasses function}
\usage{
labelPasses(blockList, startPt = c(0, 0), lastPass = 0, passDist = 5,
  ...)
}
\arguments{
\item{blockList}{[value]}

\item{startPt}{[value]. Default is "ft"}

\item{lastPass}{[value]. Default is NULL}

\item{passDist}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to label ranges using fieldBlocks.
}
\details{
[fill in details here]
}
\examples{
# none
}
