% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelRanges.R
\name{labelRanges}
\alias{labelRanges}
\title{labelRanges function}
\usage{
labelRanges(
  blockList,
  startPt = c(0, 0),
  lastRange = 0,
  rangeDist = 16,
  addRange = 0,
  ...
)
}
\arguments{
\item{blockList}{[value]}

\item{startPt}{[value]. Default is "ft"}

\item{lastRange}{[value]. Default is NULL}

\item{rangeDist}{[value]. Default is NULL}

\item{...}{additional arguments passed to text()}
}
\value{
[value]
}
\description{
function to label ranges using fieldBlocks.
}
\details{
[fill in details here]
}
\examples{
# none
}
