% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBlock.R
\name{makeBlock}
\alias{makeBlock}
\title{makeBlock function}
\usage{
makeBlock(l, blockName, pile = TRUE, serpentinePlot = FALSE, startSerp = FALSE)
}
\arguments{
\item{l}{[value]}

\item{blockName}{[value]}

\item{pile}{[value]. Default is TRUE}

\item{serpentinePlot}{[value]. Default is FALSE}

\item{startSerp}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
