% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBreedbaseTrial.R
\name{makeBreedbaseTrial}
\alias{makeBreedbaseTrial}
\title{makeBreedbaseTrial function to create breedbase multi trial upload files}
\usage{
makeBreedbaseTrial(
  trial,
  bags = NULL,
  breeding_program = "Virginia Tech",
  trial_type = "phenotyping_trial",
  design_type = "RCBD",
  plantingDate = NULL,
  harvestDate = NULL,
  plot_width = NULL,
  plot_length = NULL,
  checks = NULL,
  field_size = NULL
)
}
\arguments{
\item{trial}{object of class trialDesign}

\item{bags}{data.frame from a bag file with plot, range and pass information}

\item{breeding_program}{character string of length 1, with breeding program name}

\item{trial_type}{character string of length 1, with trial type (refer to breedbase for options)}

\item{design_type}{character string of length 1, with design type (refer to breedbase for options)}

\item{plantingDate}{optional, of the form 2020-10-23}

\item{harvestDate}{optional, of the form 2021-6-12}

\item{plot_width}{optional, numeric value for plot width}

\item{plot_length}{optional, numeric value for plot length}

\item{checks}{optional, character vector of check names}

\item{field_size}{numeric, field size in hectares}
}
\value{
data.frame formated breedbase multi trial upload file
}
\description{
function to create multi trial upload files for breedbase
}
\details{
uses T3 rules to reformat line names
}
\examples{
# none.
}
