% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFill.R
\name{makeFill}
\alias{makeFill}
\title{makePlots function
function to make fill plots. Wrapper for makePlots where all plots are fill.}
\usage{
makeFill(n, ...)
}
\arguments{
\item{n}{integer vector to determine number of fill plots}

\item{....}{Arguments passes to makePlots}
}
\value{
object of class fieldPLots
}
\description{
makePlots function
function to make fill plots. Wrapper for makePlots where all plots are fill.
}
\details{
[fill in details here]
}
\examples{
none
}
