% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHarvMasterFile.R
\name{makeHarvMasterFile}
\alias{makeHarvMasterFile}
\title{makeHarvMasterFile function}
\usage{
makeHarvMasterFile(block, printPlotNo = TRUE, rmBorderPasses = TRUE,
  rmBorderRanges = FALSE)
}
\arguments{
\item{block}{[value] Object of class 'fieldBlock'}

\item{printPlotNo}{[value]. Should plot number be printed? Default is TRUE}

\item{rmBorderPasses}{[value]. Should all border passes be removed? Default is FALSE}

\item{rmBorderRanges}{[value]. Should all border ranges be removed? Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
