% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeadrowBook.R
\name{makeHeadrowBook}
\alias{makeHeadrowBook}
\title{makeHeadrowTrays function}
\usage{
makeHeadrowBook(
  trayName,
  pop,
  ped,
  checks,
  nTrayPerPop = 3,
  trayL = 80,
  checkSetSize = 4,
  frontBack = FALSE
)
}
\arguments{
\item{trayName}{vector for the tray name}

\item{pop}{vector of population names}

\item{ped}{vector of pedigrees}

\item{checks}{vector of check varieties}

\item{nTrayPerPop}{number of trays per population, default is  3}

\item{trayL}{tray length, default is  80}

\item{checkSetSize}{number of checks per set, default is  4}

\item{frontBack}{checks at front and back of each populationo, default is  TRUE}
}
\value{
[value]
}
\description{
function to make data.frame of headrows from a list of trays and checks.
}
\details{
[fill in details here]
}
\examples{
# none
}
