% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlots.R
\name{makePlots}
\alias{makePlots}
\title{makePlots function}
\usage{
makePlots(trial, ranges, passes, rangeDist = 16, passDist = 5,
  rstart = 1, pstart = 1, angle = 0, plotNo = 1000,
  blockSize = ranges * passes, nBlock = 1, serpentine = TRUE,
  startSerp = FALSE, transpose = FALSE, ref = c(0, 0), border = 0,
  borderName = "B", borderPlotNo = 1, updateB = TRUE,
  ignorePasses = NULL, fieldViewMatrix = TRUE, fill = FALSE,
  fillBlock = FALSE, Line = NULL, Pedigree = NULL, Entry = NULL)
}
\arguments{
\item{trial}{[value]}

\item{ranges}{[value]}

\item{passes}{[value]}

\item{rangeDist}{[value]}

\item{passDist}{[value]}

\item{rstart}{[value]. Default is 1}

\item{pstart}{[value]. Default is 1}

\item{angle}{[value]. Default is 0}

\item{plotNo}{[value]. Default is 1000}

\item{blockSize}{[value]. Default is ranges * passes}

\item{nBlock}{[value]. Default is 1}

\item{serpentine}{[value]. Default is TRUE}

\item{startSerp}{[value]. Default is FALSE}

\item{transpose}{[value]. Default is FALSE}

\item{ref}{[value]. Default is c(0,0)}

\item{border}{[value]. Default is 0}

\item{borderName}{[value]. Default is "B"}

\item{borderPlotNo}{[value]. Default is 1}

\item{updateB}{[value]. Default is TRUE}

\item{ignorePasses}{[value]. Default is NULL}

\item{fieldViewMatrix}{[value]. Default is TRUE}

\item{fill}{[value]. Default is FALSE}

\item{fillBlock}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
