% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSquareField.R
\name{makeSquareField}
\alias{makeSquareField}
\title{makeSquareField function}
\usage{
makeSquareField(w, l, refx = 0, refy = 0)
}
\arguments{
\item{w, }{numeric. width of field in feet}

\item{l, }{length of field in feet}
}
\value{
object of class 'fieldBoundary'
}
\description{
function to fix planting mistakes in the field
}
\details{
[fill in details here]
}
\examples{
# none
}
