% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTrays.R
\name{makeTrays}
\alias{makeTrays}
\title{makeTrays function}
\usage{
makeTrays(plots, traySize = 40, lastTray = 0, fill = c(2, 2))
}
\arguments{
\item{plots}{either a list of trial designs, a single trial design or}

\item{pile}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to make headrow trays, particularly useful for making scab nursery trays.
}
\details{
[fill in details here]
}
\examples{
# none
}
