% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTrialLabels.R
\name{makeTrialLabels}
\alias{makeTrialLabels}
\title{makeTrialLabels function}
\usage{
makeTrialLabels(
  labPos,
  shiftLab,
  labNames,
  cols,
  a = 0,
  labh = 8,
  labw = 60,
  rotateLabs = 0,
  labPlots = TRUE,
  ...
)
}
\arguments{
\item{labPos}{matrix with 2 columns indicating x and y positions of labels, or a list of vectors of length 2, with x and y coordinates per label, or a list of fieldPlots objects.}

\item{shiftLab}{list of vector of vertical adjustments for trial labels, useful when fieldplots are provided to labPos, to adjust the label position vertically}

\item{labNames}{vector of label names.}

\item{cols}{colors to be used for label names}

\item{a}{numeric. angle at which labels should be drawn. Default is 0.}

\item{labh}{numeric. label height. Default is 8}

\item{labw}{numeric. label width. Default is 60}

\item{rotateLabs}{numjeric. rotation for labels. Default is 0}

\item{labPlots}{logical. Should plots be labeled? Useful for map development when set to FALSE to not print labels. Default is TRUE}

\item{...}{additional arguments passed to text()}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
