% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixToDist.R
\name{matrixToDist}
\alias{matrixToDist}
\title{plotField function}
\usage{
matrixToDist(m, reference = NULL)
}
\arguments{
\item{m}{n x 2 matrix of points on boundary of field. the first row will be assumed to be the reference point if no argument is provided to 'reference'}
}
\value{
object of class field boundary
}
\description{
function to make fieldBoundary from matrix of points.
}
\details{
[fill in details here]
}
\examples{
none
}
