% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanAngle.R
\name{meanAngle}
\alias{meanAngle}
\title{meanAngle function}
\usage{
meanAngle(a1, a2)
}
\arguments{
\item{a1}{[value]}

\item{a2}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
