% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldBlock.R
\name{ncol.fieldBlock}
\alias{ncol.fieldBlock}
\title{ncol.fieldBlock}
\usage{
ncol.fieldBlock(x)
}
\arguments{
\item{x}{object of class fieldBlock}
}
\description{
ncol.fieldBlock
}
\details{
determine number of columns (passes) in fieldBlock
}
