% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldBlock.R
\name{nrow.fieldBlock}
\alias{nrow.fieldBlock}
\title{nrow.fieldBlock}
\usage{
nrow.fieldBlock(x)
}
\arguments{
\item{x}{object of class fieldBlock}
}
\description{
nrow.fieldBlock
}
\details{
determine number of rows (ranges) in fieldBlock
}
