% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberPasses.R
\name{numberPasses}
\alias{numberPasses}
\title{numberPasses function}
\usage{
numberPasses(startPt, ...)
}
\arguments{
\item{startPt}{[value]}
}
\value{
[value]
}
\description{
function to add pass numbers to map
}
\details{
[fill in details here]
}
\examples{
none
}
