% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotField.R
\name{plotField}
\alias{plotField}
\title{plotField function}
\usage{
plotField(
  boundary,
  offset = 5,
  units = "ft",
  hshift = 3,
  vshift = 1,
  showBoundry = TRUE,
  showPtNo = FALSE,
  sideLength = TRUE,
  ...
)
}
\arguments{
\item{boundary}{[value]}

\item{offset}{[value]. Default is 5}

\item{units}{[value]. Default is "ft"}

\item{hshift}{[value]. Default is 3}

\item{vshift}{[value]. Default is 1}

\item{showBoundry}{[value]. Default is TRUE}

\item{showPtNo}{[value]. Default is FALSE}

\item{sideLength}{[value]. Default is TRUE}

\item{...}{arguments passed to text()}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
