% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGPS.R
\name{plotGPS}
\alias{plotGPS}
\title{plotGPS function}
\usage{
plotGPS(gps, number = TRUE, a = 0)
}
\arguments{
\item{gps}{[value]}

\item{number}{[value]. Default is TRUE}

\item{a}{[value]. Default is 0}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
