% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStartAngle.R
\name{plotStartAngle}
\alias{plotStartAngle}
\title{plotStartAngle function}
\usage{
plotStartAngle(startPt, pt1, pt2, refAngle1, refAngle2, rscale = 0.7,
  offset = 4, txtSize = 1, units = "ft")
}
\arguments{
\item{startPt}{[value]}

\item{pt1}{[value]}

\item{pt2}{[value]}

\item{refAngle1}{[value]}

\item{refAngle2}{[value]}

\item{rscale}{[value]. Default is 0.7}

\item{offset}{[value]. Default is 4}

\item{txtSize}{[value]. Default is 1}

\item{units}{[value]. Default is "ft"}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
