% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointDist.R
\name{pointDist}
\alias{pointDist}
\title{pointDist function}
\usage{
pointDist(pt1, pt2)
}
\arguments{
\item{pt1}{[value]}

\item{pt2}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
