% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printBags.R
\name{printBags}
\alias{printBags}
\title{printBags function}
\usage{
printBags(plots, addPass = 0, blockName = NULL)
}
\arguments{
\item{plots}{[value]}

\item{addPass}{[value]. Default is 0}

\item{blockName}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
