% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printStakes.R
\name{printStakes}
\alias{printStakes}
\title{printStakes function}
\usage{
printStakes(plots, addPass = 0, every = 5, tooClose = 2, blockName = NULL)
}
\arguments{
\item{plots}{[value]}

\item{addPass}{[value]. Default is 0}

\item{every}{[value]. Default is 5}

\item{tooClose}{[value]. Default is 2}

\item{blockName}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
