% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radToDeg.R
\name{radToDeg}
\alias{radToDeg}
\title{radToDeg function}
\usage{
radToDeg(rad)
}
\arguments{
\item{rad}{numeric. radians. [value]}
}
\value{
[value]
}
\description{
function to convert radians to degrees.
}
\details{
[fill in details here]
}
\examples{
# none
}
