% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randoToTrialDesign.R
\name{randoToTrialDesign}
\alias{randoToTrialDesign}
\title{randoToTrialDesign function}
\usage{
randoToTrialDesign(rando, ignoreTrials = NULL, returnList = FALSE)
}
\arguments{
\item{rando}{data.frame from output of randomizeTrial}

\item{ignoreTrials}{character vector of trials to be ignored}
}
\value{
[value] object of class trialDesign
}
\description{
function to produce an object of class 'trialDesign' from the output of randomizeTrial
}
\details{
[fill in details here]
}
\examples{
# none
}
