% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGoogleGPS.R
\name{readGoogleGPS}
\alias{readGoogleGPS}
\title{readGoogleGPS function}
\usage{
readGoogleGPS(googleCSV, ptOrder = NULL, ignore = NULL)
}
\arguments{
\item{googleCSV}{[value]}

\item{ptOrder}{[value]. Default is NULL}

\item{ignore}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
