% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalcDistAngle.R
\name{recalcDistAngle}
\alias{recalcDistAngle}
\title{recalcDistAngle function}
\usage{
recalcDistAngle(boundry)
}
\arguments{
\item{boundry}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
