% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reRandomizeTrial.R
\name{reRandomizeTrial}
\alias{reRandomizeTrial}
\title{reRandomizeTrial function}
\usage{
reRandomizeTrial(
  ent,
  rando,
  test = NULL,
  year = NULL,
  loc = NULL,
  grams = NULL
)
}
\arguments{
\item{ent}{[value]}

\item{rando}{previous randomization output from randomizeTrial()}

\item{test}{[value]}

\item{year}{[value]}

\item{loc}{[value]}

\item{grams}{[value]}
}
\value{
[value] object of class fieldTrial
}
\description{
function to re-randomize trials. Useful if the grower wants to plant multiple locations were blocks can be harvested for seed in some or all locations.
}
\details{
[fill in details here]
}
\examples{
# none
}
