% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{rotate function}
\usage{
rotate(X, theta, deg = TRUE)
}
\arguments{
\item{X}{[value]}

\item{theta}{[value]}

\item{deg}{[value]. Default is TRUE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
