% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotateMap.R
\name{rotateMap}
\alias{rotateMap}
\title{rotateMap function}
\usage{
rotateMap(map, theta)
}
\arguments{
\item{map}{[value]}

\item{theta}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
