% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotateMultiMaps.R
\name{rotateMultiMaps}
\alias{rotateMultiMaps}
\title{rotateMultiMaps function}
\usage{
rotateMultiMaps(l, a)
}
\arguments{
\item{l}{[value]}

\item{a}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
