% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftPlot.R
\name{shiftPlot}
\alias{shiftPlot}
\title{shiftPlot function}
\usage{
shiftPlot(plots, plotNo, ang1, ang2 = ang1, offsetx = 0, offsety = 0,
  corner = "bottomright", plotref = FALSE, plotstart = FALSE)
}
\arguments{
\item{plots}{[value]}

\item{plotNo}{[value]}

\item{ang1}{[value]}

\item{ang2}{[value]. Default is ang1}

\item{offsetx}{[value]. Default is 0}

\item{offsety}{[value]. Default is 0}

\item{corner}{[value]. Default is "bottomright"}

\item{plotref}{[value]. Default is FALSE}

\item{plotstart}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
