% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftPt.R
\name{shiftPt}
\alias{shiftPt}
\title{shiftPt function}
\usage{
shiftPt(pt, a = 0, x = 0, y = 0, a2 = a)
}
\arguments{
\item{pt}{[value]}

\item{a}{[value]. Default is 0}

\item{x}{[value]. Default is 0}

\item{y}{[value]. Default is 0}

\item{a2}{[value]. Default is a}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
