% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackPlots.R
\name{stackPlots}
\alias{stackPlots}
\title{stackPlots function}
\usage{
stackPlots(plotList)
}
\arguments{
\item{plotList}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
