% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T3ize.R
\name{T3ize}
\alias{T3ize}
\title{T3ize function}
\usage{
T3ize(x)
}
\arguments{
\item{x}{vector of line names to be reformatted}
}
\value{
formated line names
}
\description{
function to reformat line names using T3 rules
}
\details{
uses T3 rules to reformat line names
}
\examples{
none
}
