% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takeout.R
\name{takeout}
\alias{takeout}
\title{GPStoDist function}
\usage{
takeout(takeoutFile, rowOrder = NULL, ...)
}
\arguments{
\item{takeoutFile}{path to google takeout csv file.}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
