% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdToDf.R
\name{tdToDf}
\alias{tdToDf}
\title{tdToDf function}
\usage{
tdToDf(td, vars = c("plotNo", "plotName", "block", "trialName", "Entry",
  "Line", "Pedigree"))
}
\arguments{
\item{td}{object of class 'trialDesign'}

\item{vars}{variables that should be included in the data.frame}
}
\value{
data.frame
}
\description{
function to turn an object of class 'trialDesign' into a data.frame
}
\details{
[fill in details here]
}
\examples{
# none
}
