% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transLatLon.R
\name{transLatLon}
\alias{transLatLon}
\title{transLatLon function}
\usage{
transLatLon(lat1, lon1, lat2, lon2, ref, units = "ft")
}
\arguments{
\item{lat1}{[value]}

\item{lon1}{[value]}

\item{lat2}{[value]}

\item{lon2}{[value]}

\item{ref}{[value]}

\item{units}{[value]. Default is "ft"}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
