% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trialDesign.R
\docType{class}
\name{trialDesign-class}
\alias{trialDesign-class}
\alias{trialDesign}
\title{An S4 class to represent entry}
\description{
An S4 class to represent entry
}
\section{Slots}{

\describe{
\item{\code{plotInfo}}{data.frame}
}}

