% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeStakes.R
\name{writeStakes}
\alias{writeStakes}
\title{writeStakes function}
\usage{
writeStakes(
  ...,
  dir = "",
  addPass = 0,
  every = 5,
  tooClose = 2,
  blockName = NULL
)
}
\arguments{
\item{...}{objects of class 'fieldPlots' or 'fieldBlock' to write stake files for.}

\item{dir}{[value]. Default is ""}

\item{addPass}{[value]. Default is 0}

\item{every}{[value]. Default is 5}

\item{tooClose}{[value]. Default is 2}

\item{blockName}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
