% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglePoint.R
\name{anglePoint}
\alias{anglePoint}
\title{anglePoint function}
\usage{
anglePoint(d, theta, ref = c(0, 0))
}
\arguments{
\item{d}{[value]}

\item{theta}{[value]}

\item{ref}{[value]. Default is c(0, 0)}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
