% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPal.R
\docType{data}
\name{cols10}
\alias{cols10}
\title{color pallete'10}
\format{
An object of class \code{character} of length 10.
}
\usage{
cols10
}
\description{
color pallete'10
}
\details{
color palette 10
}
\examples{
# none
}
\keyword{datasets}
