% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degToRad.R
\name{degToRad}
\alias{degToRad}
\title{degToRad function}
\usage{
degToRad(deg)
}
\arguments{
\item{deg}{numeric. degrees.}
}
\value{
numeric. radians
}
\description{
function to convert degrees to radians
}
\details{
convert degrees to radians
}
\examples{
degToRad(45) == pi / 4
}
