% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawLegend.R
\name{drawLegend}
\alias{drawLegend}
\title{drawLegend function}
\usage{
drawLegend(plotList, boundry, cols, shiftLeg = c(0, 0),
  labNames = NULL, scaleLeg = 0.33, rotateLabs = 0, ...)
}
\arguments{
\item{plotList}{[value]}

\item{boundry}{[value]}

\item{cols}{[value]}

\item{shiftLeg}{[value]}

\item{labNames}{[value]}

\item{scaleLeg}{[value]. Default is 0.33}

\item{rotateLabs}{[value]. Default is 0}
}
\value{
[value]
}
\description{
function to draw legend for field map.
}
\details{
[fill in details here]
}
\examples{
# none
}
