% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawPlots.R
\name{drawPlots}
\alias{drawPlots}
\title{drawPlots function}
\usage{
drawPlots(
  plots,
  angle = 0,
  labAngle = angle + 90,
  plotNos = TRUE,
  add = TRUE,
  ptrim = 1,
  rtrim = 2,
  color = NA,
  showBorderNum = FALSE,
  ...
)
}
\arguments{
\item{plots}{object of class 'fieldPlots'}

\item{angle}{numeric. Angle of which to draw plots}

\item{labAngle}{numeric. Angle at which to draw plot labels}

\item{plotNos}{logical. Should plot numbers be displayed? Default is TRUE}

\item{add}{logical. should plots be drawn on an existing R device? Default is TRUE}

\item{ptrim}{numeric. Distance to remove between passes when drawing plots. Example, passes may be 5 feet wide, but plots are only 4 feet wide, then ptrim should be 5-4=1. Default is 1}

\item{rtrim}{numeric. Distance to remove between passes when drawing plots. Example, passes may be 16 feet apart, but plots are then trimmed to 14 feet long, then rtrim should be 16-14=2.Default is 2}

\item{color}{character. color to be used for plots}

\item{showBorderNum}{logical. Should border numbers be shown?}

\item{...}{arguments passed to text()}
}
\value{
nothing
}
\description{
function to draw plots on an R graphics device.
}
\details{
drawPlots draws a graphical representation of field plots to an R device, such as a pdf.
}
\examples{
# none
}
