% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldPlots.R
\docType{class}
\name{fieldPlots}
\alias{fieldPlots}
\alias{rbind,fieldPlots-method}
\alias{cbind,fieldPlots-method}
\title{An S4 class to represent field plots}
\usage{
\S4method{rbind}{fieldPlots}(..., deparse.level = 1)

\S4method{cbind}{fieldPlots}(..., deparse.level = 1)
}
\description{
An S4 class to represent field plots
}
\section{Slots}{

\describe{
\item{\code{centers}}{list of centers}

\item{\code{corners}}{list of corners}

\item{\code{matrix}}{matrix of matrix}

\item{\code{needStake}}{list of needStake}

\item{\code{borderPasses}}{numeric of borderPasses}

\item{\code{trialName}}{character of trialName}
}}

