% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPStoDist.R
\name{GPStoDist}
\alias{GPStoDist}
\title{GPStoDist function}
\usage{
GPStoDist(boundary, units = "ft", refPoint = NULL)
}
\arguments{
\item{boundary}{[value]}

\item{units}{[value]. Default is "ft"}

\item{refPoint}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to  make fieldBoundary from matrix of GPS points. Returns a flattened matrix in feet (default) or meters (when units argument set to 'm')
}
\details{
[fill in details here]
}
\examples{
none
}
