% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTrialLabels.R
\name{makeTrialLabels}
\alias{makeTrialLabels}
\title{makeTrialLabels function}
\usage{
makeTrialLabels(
  labPos,
  shiftLab,
  labNames,
  cols,
  a = 0,
  labh = 8,
  labw = 60,
  rotateLabs = 0,
  labPlots = TRUE,
  ...
)
}
\arguments{
\item{labPos}{[value]}

\item{shiftLab}{[value]}

\item{labNames}{[value]}

\item{cols}{[value]}

\item{a}{[value]. Default is 0}

\item{labh}{[value]. Default is 8}

\item{labw}{[value]. Default is 60}

\item{rotateLabs}{[value]. Default is 0}

\item{labPlots}{[value]. Default is TRUE}

\item{...}{additional arguments passed to text()}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
