% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printBags.R
\name{printBags}
\alias{printBags}
\title{printBags function}
\usage{
printBags(plots, addPass = 0, countBorder = TRUE, blockName = NULL,
  printBorder = FALSE)
}
\arguments{
\item{plots}{[value]}

\item{addPass}{numeric. number of passes to add. Default is 0}

\item{countBorder}{logical. Should the border passes be counted? Default is TRUE}

\item{blockName}{chracter. name of block. Default is NULL}

\item{printBorder}{logical. Should bags for the border passes be printed? Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
