% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radToDeg.R
\name{radToDeg}
\alias{radToDeg}
\title{radToDeg function}
\usage{
radToDeg(rad)
}
\arguments{
\item{radToDeg}{<- function (rad [value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
