% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizeTrial.R
\name{randomizeTrial}
\alias{randomizeTrial}
\title{randomizeTrial function}
\usage{
randomizeTrial(ent, test, year, loc, reps, grams, randomize = "RCBD",
  otherCols = NULL)
}
\arguments{
\item{ent}{[value]}

\item{year}{[value]}

\item{loc}{[value]}

\item{reps}{[value]}

\item{grams}{[value]}

\item{randomize}{[value]. Default is 'RCBD'}

\item{testName}{[value]}
}
\value{
[value] object of class fieldTrial
}
\description{
function to randomize trials
}
\details{
[fill in details here]
}
\examples{
none
}
