% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBags.R
\name{writeBags}
\alias{writeBags}
\title{writeBags function}
\usage{
writeBags(..., dir = "", addPass = 0, countBorder = TRUE,
  blockName = NULL)
}
\arguments{
\item{dir}{[value]. Default is ""}

\item{addPass}{[value]. Default is 0}

\item{countBorder}{logical. SHould the border be counted? Default is TRUE}

\item{blockName}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
