% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHarvMasterFile.R
\name{writeHarvMasterFile}
\alias{writeHarvMasterFile}
\title{writeHarvMasterFile function}
\usage{
writeHarvMasterFile(block, dir, ...)
}
\arguments{
\item{block}{[value]. object of class 'fieldBlock'}

\item{dir}{[value]. Default is ""}

\item{...}{additional arguments passed to write.csv()}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
